add fizz buzz snippet
"fizz buzz" was a short technical competency test that used to do the
rounds in interviews a while back: write a function that, given a list
of numbers, will iterate through them and if the number is divisible by
three replace it with the word "fizz", if the number is divisible by
five with the word "buzz", and if the number is divisible by both three
and five with the word "fizzbuzz":

    perl -E 'say $_ % 3 == 0 && $_ % 5 == 0 ? "fizzbuzz" : $_ % 3 == 0 ?
    "fizz" : $_ % 5 == 0 ? "buzz" : $_ for 1 .. 30'
    1
    2
    fizz
    4
    buzz
    fizz
    7
    8
    fizz
    buzz
    11
    fizz
    13
    14
    fizzbuzz
    16
    17
    fizz
    19
    buzz
    fizz
    22
    23
    fizz
    buzz
    26
    fizz
    28
    29
    fizzbuzz

the "technical" competency here is just a probe for: does this person
know about loops, conditionals, and the modulus operator. it's a pretty
poor probe for "is this person a competent software engineer"

the solution above is mine, that just took me about 30s to write. the
added snippet here is from an AI system prompted to generate the code.
it's wrong on so many levels, but that's not surprising given this
"challenge" is so well known that there is inevitably a huge corpus of
badly coded and completely incorrect examples out there for it to pull
from.

the snippet added here:

Doesn't work (it doesn't solve the problem).
    * `@factors`? Where is that populated? Global i guess?
    * `$starting` and `$ending` also? Are these supposed to be globals too?
    * "Check if current number is divisible by... itself" - Did someone fail basic math?

It's not idiomatic.
    * 30 lines to do what can be done (clearly) in 5.
    * or even as a one liner if you're throwing it away immediately

It's syntactically dubious.
    * WTF is `%self` doing in there?
    * In fact, why are we assigning single values into an associated array (a hash)?

It doesn't even compile.
    * Look, I know humans are not compilers and we can elide or infer details, but
    * `foreach ... else`? That's not a thing. Sure, it would actually be quite a nice feature...
    * But it's currently not code that will compile under even a modern perl.

so it's a nice example - it was given in a talk at a Perl conference
about a recently created system to use an LLM to generate Perl code and
how that might be useful. i think this snippet proves quite the opposite